Add auto-insert-mode configs and files

+++ install.sh
@@ -5,6 +5,13 @@ if [ ! -e "$HOME/.emacs.d" ]; then
   mkdir "$HOME/.emacs.d"
 fi
 
-echo "make a symbolic link to dotfiles/init.el"
-ln -s "$(pwd)/init.el" "$HOME/.emacs.d/init.el"
 
+if [ ! -e "$HOME/.emacs.d/init.el" ]; then
+  echo "make a symbolic link to dotfiles/init.el"
+  ln -s "$(pwd)/init.el" "$HOME/.emacs.d/init.el"
+fi
+
+if [ ! -e "$HOME/.emacs.d/insert" ]; then
+  echo "make a symbolic link to dotfiles/insert"
+  ln -s "$(pwd)/insert" "$HOME/.emacs.d/insert"
+fi
